Remove old data types file Add missing data interfaces in the right data types file Add ID in creation response interface Add game and game creation response interfaces
